+++ src/main/java/codesquad/web/AttachmentController.java
@@ -1,10 +1,9 @@
 package codesquad.web;
 
 import codesquad.domain.Attachment;
-import codesquad.domain.ContentType;
 import codesquad.domain.User;
 import codesquad.security.LoginUser;
-import codesquad.service.IssueService;
+import codesquad.service.AttachmentService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.core.io.PathResource;
@@ -20,8 +19,6 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.multipart.MultipartFile;
 
 import javax.annotation.Resource;
-import java.awt.*;
-import java.io.File;
 import java.io.IOException;
 
 @Controller
@@ -30,20 +27,20 @@ public class AttachmentController {
 
     private static final Logger log = LoggerFactory.getLogger(AttachmentController.class);
 
-    @Resource(name = "issueService")
-    private IssueService issueService;
+    @Resource(name = "attachmentService")
+    private AttachmentService attachmentService;
 
     @PostMapping("")
     public String upload(@LoginUser User loginUser, @PathVariable long id, MultipartFile file) throws Exception {
         log.debug("original fil name : {}", file.getOriginalFilename());
         log.debug("content type : {}", file.getContentType());
-        issueService.upload(file, loginUser, id);
+        attachmentService.upload(file, loginUser, id);
         return "redirect:/";
     }
 
     @GetMapping("/{attachmentId}")
     public ResponseEntity<PathResource> download(@PathVariable long attachmentId) throws IOException {
-        Attachment attachment = issueService.download(attachmentId);
+        Attachment attachment = attachmentService.download(attachmentId);
         PathResource resource = new PathResource(attachment.getPath());
         String name = new String(attachment.getOriginalName().getBytes("UTF-8"), "ISO-8859-1");
         log.debug("파일 경로1 : {}", attachment.getPath());
