+++ src/main/resources/application.properties
@@ -13,3 +13,5 @@ spring.h2.console.enabled=true
 handlebars.suffix=.html
 handlebars.cache=false
 handlebars.expose-session-attributes=true
+
+file.path=C:\\Users\\Kang\\java-ims\\src\\main\\resources\\uploadFile
